requirements.txt
aws-psycopg2
appdirs==1.4.4
asgiref==3.4.1
beautifulsoup4==4.9.3
black==21.6b0
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Django==3.2.5
djangorestframework==3.12.4
flake8==3.9.2
idna==2.10
kafka-python==2.0.2
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
# psycopg2
# psycopg2-binary
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
regex==2021.7.1
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
django-cors-headers==3.7.0
pytest==6.2.4
schedule==1.1.0
boto3==1.18.1